Implement AbbrebiationTranslator.translate_string()
