+++ README.md
@@ -45,3 +45,4 @@ changes.
 ## TODO
 
 - Fix func args in `test*.py`
+- Move majority of moudules to a subdir
